Paint remixer theatres black.
