+++ packages/remixer/src/Source.js
@@ -2,13 +2,16 @@ import React, { useRef, useState, useEffect, useCallback, useMemo } from 'react'
 
 import Box from '@mui/material/Box';
 import Container from '@mui/material/Container';
+import Toolbar from '@mui/material/Toolbar';
 import { styled } from '@mui/material/styles';
 
 import { Stage, Transcript } from './components';
+import SourceTopbar from './components/SourceTopbar';
 
 const PREFIX = 'Source';
 const classes = {
   root: `${PREFIX}-root`,
+  toolbar: `${PREFIX}-toolbar`,
   theatre: `${PREFIX}-theatre`,
   transcript: `${PREFIX}-transcript`,
 };
@@ -20,8 +23,16 @@ const Root = styled('div')(({ theme }) => ({
   height: '100%',
   justifyContent: 'space-between',
   position: 'relative',
+  [`& .${classes.toolbar}`]: {
+    backgroundColor: 'black',
+    color: theme.palette.primary.contrastText,
+    left: 0,
+    position: 'absolute',
+    right: 0,
+    top: 0,
+    '& .Mui-disabled': { color: 'rgba(255,255,255,0.5) !important' },
+  },
   [`& .${classes.theatre}`]: {
-    aligenItems: 'center',
     backgroundColor: 'black',
     color: theme.palette.primary.contrastText,
     display: 'flex',
@@ -67,6 +78,14 @@ const Source = props => {
 
   return (
     <Root className={classes.root}>
+      {!isSingleMedia && (
+        <>
+          <Toolbar sx={{ bgcolor: 'black', width: '100%' }} />
+          <Box className={classes.toolbar}>
+            <SourceTopbar {...props} />
+          </Box>
+        </>
+      )}
       <Box className={classes.theatre}>
         <Container maxWidth="sm" sx={{ px: { xs: 0, sm: 3 } }}>
           <Stage {...{ blocks, media, players, reference, time, setTime, singlePlayer, singlePlayerOffset }} />
